+++ components/ActivityCard.tsx
