+++ components/layout/AppFooter.tsx
